change over to registered property

+++ src/engraving/style/styledef.cpp
@@ -31,7 +31,6 @@
 #include "dom/realizedharmony.h"
 #include "dom/stafftype.h"
 #include "dom/textbase.h"
-#include "dom/timesig.h"
 #include "dom/tuplet.h"
 #include "dom/types.h"
 
@@ -470,7 +469,7 @@ const std::array<StyleDef::StyleValue, size_t(Sid::STYLES)> StyleDef::styleValue
 
     styleDef(multiVoiceRestTwoSpaceOffset,               false),
     styleDef(mergeMatchingRests,                         false),
-    styleDef(fullMeasureRestType,                        PropertyValue(static_cast<int>(TimeSigFullMeasureRestType::WHOLE_OR_BREVE))),
+    styleDef(fullMeasureRestType,                        PropertyValue(TimeSigFullMeasureRestType::WHOLE_OR_BREVE)),
     styleDef(hideFullMeasureRestsWhenEmpty,              false),
     styleDef(createMultiMeasureRests,                    false),
     styleDef(minEmptyMeasures,                           PropertyValue(1)),
